+++ lib/widgets/journal/text_viewer_widget.dart
@@ -1,9 +1,7 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_quill/flutter_quill.dart' hide Text;
+import 'package:flutter/cupertino.dart';
+import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:lotti/classes/entry_text.dart';
 import 'package:lotti/themes/theme.dart';
-import 'package:lotti/widgets/journal/editor/editor_styles.dart';
-import 'package:lotti/widgets/journal/editor/editor_tools.dart';
 
 class TextViewerWidget extends StatelessWidget {
   const TextViewerWidget({
@@ -13,32 +11,24 @@ class TextViewerWidget extends StatelessWidget {
   });
 
   final EntryText? entryText;
-  final double? maxHeight;
+  final double maxHeight;
 
   @override
   Widget build(BuildContext context) {
-    if (entryText == null || entryText?.plainText == '\n') {
-      return const SizedBox.shrink();
-    }
-
-    final controller = makeController(serializedQuill: entryText?.quill);
-
-    return IgnorePointer(
-      child: SingleChildScrollView(
-        child: QuillEditor(
-          controller: controller,
-          readOnly: true,
-          scrollController: ScrollController(),
-          scrollable: true,
-          focusNode: FocusNode(canRequestFocus: false),
-          autoFocus: false,
-          expands: false,
-          maxHeight: maxHeight,
-          padding: const EdgeInsets.only(top: 8, bottom: 16),
-          keyboardAppearance: keyboardAppearance(),
-          customStyles: customTextViewerStyles(
-            textColor: styleConfig().primaryTextColor,
-            codeBlockBackground: styleConfig().primaryColorLight,
+    return LimitedBox(
+      maxHeight: maxHeight,
+      child: Markdown(
+        data: entryText?.markdown ?? '',
+        shrinkWrap: true,
+        styleSheet: MarkdownStyleSheet.fromCupertinoTheme(
+          CupertinoThemeData(
+            textTheme: CupertinoTextThemeData(
+              textStyle: TextStyle(
+                color: styleConfig().primaryTextColor,
+                fontSize: fontSizeMedium,
+                fontFamily: mainFont,
+              ),
+            ),
           ),
         ),
       ),