+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed:
+- Replace read-only flutter_quill with flutter_markdown for better scroll performance
+
+### Fixed:
+- Journal card text color when using bright theme
+
+## [0.8.257] - 2023-02-06
 ### Fixed:
 - Locking issue in sync